requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aioredis==2.0.1
aioschedule==0.5.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
DateTime==5.5
dnspython==2.7.0
dotenv-python==0.0.1
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.8
frozenlist==1.5.0
gigachat==0.1.39.post1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
magic-filter==1.0.12
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
redis==5.2.1
requests==2.32.3
rsa==4.9
selenium==4.30.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wheel==0.45.1
win32_setctime==1.2.0
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
